Proposal to remove Identifier model
Deposit: if we know the repository, do we need an Identifer object?
Grant: renamed to clarify that this is a local id, this is used in other
metadata standards. We could assume there will only be one and internally
we would know what it would represent. Removed alternateIds.
Journal: can we just assign multiple string issns? is it necessary to
know epub issn etc?

remove altids