Adds proposed Repository and Policy objects to model
Initial proposal to flesh out policy/repository part of model to support Ember workflow.  Converts Repository string to an object and links it to funder through Policy.  There are potential ways to simplify, for example, could have 1-1 Funder-Policy-Repository links. The fields are based on the demo interface that we have, which displays Policy and Repository information, with no corresponding data model. Ember logic could select the Policy with type "insitution" as a default. Policy model could be expanded to incorporate submission type e.g. data in the future.