Replace Submission.authors with corresponding author info
To avoid muddy People records, rather than creating a new Person object
for each author, we are only capturing the corresponding author name and
email.  Updated the diagram to show that a Submission can now be
attached to one and only one person, though the form of this link is
currently a question mark. See issue #14.